% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutatr.R
\name{generate_mutations}
\alias{generate_mutations}
\title{Generate n mutations for the given abstract syntax tree.}
\usage{
generate_mutations(asts, n)
}
\arguments{
\item{asts}{The abstract syntax trees to generate mutations for. Must be a named
lists with the name being the file path and the value being the abstract syntax tree.}

\item{n}{The number of mutations to generate.}
}
\value{
A list of n mutated abstract syntax trees with the applied mutation
}
\description{
Generate n mutations for the given abstract syntax tree.
}
