\name{object_scope}
\alias{object_scope}
\title{Add object scope to a function...}
\usage{object_scope(res, self)}

\description{
  Add object scope to a function
}
\keyword{internal}
\arguments{
  \item{res}{object to add object scope to (does nothing if not a function)}
  \item{self}{self object}
}
