% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutatr.R
\name{generate_mutants}
\alias{generate_mutants}
\title{Generate n mutations for the given abstract syntax tree.}
\usage{
generate_mutants(
  files,
  n,
  probabilities = list(),
  seed = NULL,
  filters = list(file = function(f) TRUE, srcref = function(f, s) TRUE)
)
}
\arguments{
\item{files}{Set of files to generate mutations for.}

\item{n}{The number of mutations to generate.}

\item{probabilities}{A named list of probabilities for each mutation. If a mutation
is not in the list, the default probability is used.}

\item{seed}{The seed that determines what mutations are selected. If NULL, a random
seed is used.}

\item{filters}{A list of two filters. \code{filters$file} takes a file name and returns whether
the package should consider this file. \code{filters$srcref} takes a file and a srcref and returns
whether a mutations should be applied at this location.}
}
\value{
A list of n mutants and the used seed
}
\description{
Generate n mutations for the given abstract syntax tree.
}
