% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutatr.R
\name{generate_mutations}
\alias{generate_mutations}
\title{Generate n mutations for the given abstract syntax tree.}
\usage{
generate_mutations(ast, n)
}
\arguments{
\item{ast}{The abstract syntax tree to generate mutations for.}

\item{n}{The number of mutations to generate.}
}
\value{
A list of n mutated abstract syntax trees with the applied mutation
}
\description{
Generate n mutations for the given abstract syntax tree.
}
